Add disabled State to List View pattern
I use a light gray here for disabled rows to distinguish from the new
selected row highlighting (as it was mentioned in Github that the two
colors we have implemented are too close to tell from each other). When
hovering on it, the cursor will change to a Cannot Click State.