Update PULL_REQUEST_TEMPLATE.md
Removed introductory text, will see if this is need and if so, if it makes more sense to have this in a response comment that reflects this sentiment:

Thanks for all the great work you have done for this design contribution!  
Once the design for the pattern is submitted with a PR (Pull Request), we will do our best to get back to you, as soon as possible, so that we can get your contribution posted to the PatternFly website.